# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "NimeStream API documentation",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/report-episode": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "report_error"
                ],
                "summary": "Get all report_error",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search term",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Type Status",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "report_error"
                ],
                "summary": "Create a new report_error",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_report_error_request.CreateReportError"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/report-episode/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "report_error"
                ],
                "summary": "Get a report_error by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ReportError ID (uint)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "report_error"
                ],
                "summary": "Update a report_error",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ReportError ID (uint)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_report_error_request.UpdateReportError"
                        }
                    }
                ],
                "responses": {}
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "report_error"
                ],
                "summary": "Delete a report_error",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ReportError ID (uint)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/request-vip": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "request_vip"
                ],
                "summary": "Get all request_vip",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search term",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_util_response.SuccessWithPaginate-github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_request_vip_response_RequestVipResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "request_vip"
                ],
                "summary": "Create a new request_vip",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Payment method",
                        "name": "payment_method",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Nama pengirim",
                        "name": "atas_nama_tf",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Email",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Bukti transfer (image)",
                        "name": "bukti_tf",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/request-vip/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "request_vip"
                ],
                "summary": "Get a request_vip by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "RequestVip ID (uint)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "request_vip"
                ],
                "summary": "Update a request_vip",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "RequestVip ID (uint)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_request_vip_request.UpdateRequestVip"
                        }
                    }
                ],
                "responses": {}
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "request_vip"
                ],
                "summary": "Delete a request_vip",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "RequestVip ID (uint)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_report_error_request.CreateReportError": {
            "type": "object",
            "properties": {
                "episode_id": {
                    "type": "string"
                },
                "movie_type": {
                    "type": "string"
                },
                "problem_desc": {
                    "type": "string"
                }
            }
        },
        "github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_report_error_request.UpdateReportError": {
            "type": "object",
            "properties": {
                "movie_type": {
                    "type": "string"
                },
                "problem_desc": {
                    "type": "string"
                },
                "status_report": {
                    "type": "string"
                }
            }
        },
...
        "github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_request_vip_request.UpdateRequestVip": {
            "type": "object",
            "properties": {
                "atas_nama_tf": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "status_acc": {
                    "type": "string"
                }
            }
        },
        "github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_request_vip_response.RequestVipResponse": {
            "type": "object",
            "properties": {
                "atas_nama_tf": {
                    "type": "string"
                },
                "bukti_tf": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "payment_method": {
                    "type": "string"
                },
                "status_acc": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_util_response.SuccessWithPaginate-github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_request_vip_response_RequestVipResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_request_vip_response.RequestVipResponse"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "total_pages": {
                    "type": "integer"
                },
                "total_results": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Example Value: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}